backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
bcrypt==4.1.2
pandas==2.1.4
openpyxl==3.1.2
Werkzeug==3.0.1
marshmallow==3.20.2
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
APScheduler==3.10.4
schedule==1.2.0
